requirements.txt
# got using pipreqs
chromadb==0.4.21
Django==5.0.1
djangorestframework==3.14.0
docx==0.2.4
fitz==0.0.1.dev2
ftfy==6.1.3
openai==0.27.7
pandas==1.5.3
pdf2image==1.16.3
Pillow==9.4.0
Pillow==10.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.0
python_pptx==0.6.23
Requests==2.31.0
tabula==1.0.5
